Created Game Over Menu
Game Over Menu Scene has been added.  To test/show that it works I made a pawn that when you press the "p" key it will load the Game Over Menu.  I changed the Game Mode default pawn to this pawn so we will need to change this to the Player pawn when we can get him to be added successfully.